+++ src/drawer.canvasentry.js
@@ -61,6 +61,32 @@ export default class CanvasEntry {
                 : 'canvasentry_'
         );
         /**
+         * Whether or not the render process completed for this entry.
+         *
+         * @type {boolean}
+         */
+        this.renderComplete = false;
+        /**
+         * Cached coordinates that can be used to redraw this entry at a later
+         * time.
+         *
+         * @type {object}
+         */
+        this.cachedCoordinates = null;
+        /**
+         * Whether or not this entry is currently intersecting in the
+         * view-port.
+         *
+         * @type {boolean}
+         */
+        this.intersecting = false;
+        /**
+         * Whether or not the intersection observer API is enabled.
+         *
+         * @type {boolean}
+         */
+        this.usesIntersectionObserver = false;
+        /**
          * Canvas 2d context attributes
          *
          * @type {object}
@@ -268,12 +294,33 @@ export default class CanvasEntry {
      * @param {number} absmax Maximum peak value (absolute)
      * @param {number} halfH Half the height of the waveform
      * @param {number} offsetY Offset to the top
-     * @param {number} start The x-offset of the beginning of the area that
-     * should be rendered
+     * @param {number} start The x-offset of the beginning of the area
+     * that should be rendered
      * @param {number} end The x-offset of the end of the area that
      * should be rendered
      */
     drawLines(peaks, absmax, halfH, offsetY, start, end) {
+        if (this.usesIntersectionObserver) {
+            // do not draw unconditionally, only draw when this entry is
+            // visible in the wrapper's viewport. save coordinates until
+            // that happens so it can be redrawn later.
+            if (this.cachedCoordinates === null) {
+                this.cachedCoordinates = {
+                    peaks: peaks,
+                    absmax: absmax,
+                    halfH: halfH,
+                    offsetY: offsetY,
+                    start: start,
+                    end: end
+                };
+            }
+            if (!this.intersecting) {
+                // not visible yet
+                return;
+            }
+        }
+
+        // wave
         this.drawLineToContext(
             this.waveCtx,
             peaks,
@@ -283,7 +330,7 @@ export default class CanvasEntry {
             start,
             end
         );
-
+        // progress
         if (this.hasProgressCanvas) {
             this.drawLineToContext(
                 this.progressCtx,
@@ -295,6 +342,8 @@ export default class CanvasEntry {
                 end
             );
         }
+
+        this.renderComplete = true;
     }
 
     /**
@@ -365,6 +414,23 @@ export default class CanvasEntry {
     }
 
     /**
+     * Draw lines using cached coordinates.
+     */
+    drawLinesFromCache() {
+        // render once
+        if (!this.renderComplete && this.cachedCoordinates) {
+            this.drawLines(
+                this.cachedCoordinates.peaks,
+                this.cachedCoordinates.absmax,
+                this.cachedCoordinates.halfH,
+                this.cachedCoordinates.offsetY,
+                this.cachedCoordinates.start,
+                this.cachedCoordinates.end
+            );
+        }
+    }
+
+    /**
      * Destroys this entry
      */
     destroy() {
@@ -373,6 +439,10 @@ export default class CanvasEntry {
 
         this.progressCtx = null;
         this.progress = null;
+
+        this.renderComplete = false;
+        this.intersecting = false;
+        this.cachedCoordinates = null;
     }
 
     /**